more mingw32 patches, still not working
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@4239 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ include/qore/intern/glob.h
@@ -37,7 +37,7 @@ protected:
    names_t names;
 
 public:
-   int gl_pathc;
+   size_t gl_pathc;
    const char **gl_pathv;
 
    DLLLOCAL QoreGlobWin() : gl_pathc(0), gl_pathv(0) {
